+++ src/cb_mcp/utils/connection_string.py
@@ -1,6 +1,13 @@
-"""Helpers for parsing Couchbase connection strings."""
+"""Helpers for parsing Couchbase connection strings and deriving REST-call settings from them."""
 
+import logging
+import os
+from importlib.resources import files
 from urllib.parse import urlparse
+
+from .constants import MCP_SERVER_NAME
+
+logger = logging.getLogger(f"{MCP_SERVER_NAME}.utils.connection_string")
 
 
 def extract_hosts_from_connection_string(connection_string: str) -> list[str]:
@@ -25,3 +32,75 @@ def is_capella_connection(connection_string: str) -> bool:
     return bool(hosts) and all(
         host.lower().endswith(".cloud.couchbase.com") for host in hosts
     )
+
+
+def _get_capella_root_ca_path() -> str:
+    """Get the path to the Capella root CA certificate.
+
+    Uses importlib.resources to locate the certificate file, which works when the package is installed with fallback for development.
+
+    Returns:
+        Path to the Capella root CA certificate file.
+    """
+    try:
+        # Use importlib.resources to get the certificate path (works for installed packages)
+        cert_file = files("cb_mcp.certs").joinpath("capella_root_ca.pem")
+        # Convert to string path - this works for both installed packages and dev mode
+        return str(cert_file)
+    except (ImportError, FileNotFoundError, TypeError):
+        # Fallback for development: use src/certs/ directory
+        utils_dir = os.path.dirname(os.path.abspath(__file__))
+        src_dir = os.path.dirname(utils_dir)
+        fallback_path = os.path.join(src_dir, "certs", "capella_root_ca.pem")
+
+        if os.path.exists(fallback_path):
+            logger.info(f"Using fallback certificate path: {fallback_path}")
+            return fallback_path
+
+        # If we still can't find it, log a warning and return the fallback path anyway
+        logger.warning(
+            f"Could not locate Capella root CA certificate at {fallback_path}. "
+            "SSL verification may fail for Capella connections."
+        )
+        return fallback_path
+
+
+def determine_ssl_verification(
+    connection_string: str, ca_cert_path: str | None
+) -> bool | str:
+    """Determine SSL verification setting based on connection string and cert path.
+
+    Args:
+        connection_string: Couchbase connection string
+        ca_cert_path: Optional path to CA certificate
+
+    Returns:
+        SSL verification setting (bool or path to cert file)
+    """
+    is_tls_enabled = connection_string.lower().startswith("couchbases://")
+
+    # Priority 1: Capella connections always use Capella root CA
+    if is_capella_connection(connection_string):
+        capella_ca = _get_capella_root_ca_path()
+        if os.path.exists(capella_ca):
+            logger.info(
+                f"Capella connection detected, using Capella root CA: {capella_ca}"
+            )
+            return capella_ca
+        logger.warning(
+            f"Capella CA certificate not found at {capella_ca}, "
+            "falling back to system CA bundle"
+        )
+        return True
+
+    # Priority 2: Non-Capella TLS connections use provided cert or system CA bundle
+    if is_tls_enabled:
+        if ca_cert_path:
+            logger.info(f"Using provided CA certificate: {ca_cert_path}")
+            return ca_cert_path
+        logger.info("Using system CA bundle for SSL verification")
+        return True
+
+    # Priority 3: Non-TLS connections (HTTP), disable SSL verification
+    logger.info("Non-TLS connection, SSL verification disabled")
+    return False
